requirements.txt
facker 
azure-eventhub 
azure-identity
dotenv
aiohttp